move lógica de deduplicacao de pipeline para raiz
